requirement.txt
ujson
fastapi
fastapi-cli
pydantic-settings
groq
langchain
langchain-groq
